Commit wordpress site, database info is removed
Decided to migrate to wordpresss from our originate code base. Wordpress offers an in-house admin suite, and user registration. In the effort of not attempting to reinvent the wheel we found this to be the best solution to deliver a working project that hit most of the requirements. Not only that but as wordpress uses a database we can easily implement the score tracking into it and plan on doing so in a later commit.